Replace tuple of options with BlockLocation
The internal method get_block_from_main_cache_or_blockstore_name
returned a tuple of options, which is both unwieldy and has some very
specific mutually exclusive requirements, in practice.

Replacing this an emum, that describes the three possible outcomes in
code makes it easier to consume cognatively and programmatically.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>