Make BlockManager threadsafe
This commit splits the BlockManager into BlockManagerState which has
the fields that were previously on the BlockManager and a new threadsafe
BlockManager with a single internal field of
Arc<RwLock<BlockManagerState>>. This affects how the `get`, `branch`,
and `branch_diff` Iterators work. They now take the
Arc<RwLock<BlockManagerState>> instead of &BlockManager.

Signed-off-by: Boyd Johnson <bjohnson@bitwise.io>