Reference blocks before validating
The block validator requires that predecessors stay in the block
manager, so the chain controller should reference the block before
sending it off for validation.  Once it receives the result, it
decrements the reference.

This fixes an issue that may occur with longer forks that are aged out
of the ForkPool, while a block is in the schedule to be validated.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>