Only acquire write lock when completing is likely
Rather than trying to hold an external reference to the previous block while
trying to complete a block, which requires to BlockManager write lock
acquisitions, delay the write lock acquisition until we are trying to put the
new block into the block manager.

If the predecessor is missing when we try to put the new block, treat the block
as incomplete and request the predecessor.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>