Modify devmode-liveness min and max wait timers
One of the checks in devmode-liveness is that 100 batches are in the
chain. Prior to this commit the min wait time is 1 and the max is 3. With
those numbers there is a high likelihood of not having 100 batches in
the chain by the time the check happens. This
commit changes min to 2 and max to 4.

There are several ways to solve the problem:
   - change the min and max wait timer: this solution
   - change the test to require fewer blocks in the chain
   - change the workload to send more batches per second

Signed-off-by: Boyd Johnson <bjohnson@bitwise.io>