Undo include (move content from separate file into topic)
Using the include directive to add contents from a separate file
causes a problem with the search index.

Searching for any content in this file will find a topic called <no title>.
When viewed, this untitled topic isn't "in" the doc nav - it doesn't have
any Previous and Next buttons on the page.

This PR simply moves the relevant content into the topic.

Signed-off-by: Anne Chenette <chenette@bitwise.io>