Add queue to consensus notifier
Adds a threadsafe queue to the ConsensusNotifier class in the validator
to ensure that all messages intended for the consensus engine get sent.
Previously, messages to be sent to the engine before it was registered
were simply dropped, which would result in missed BlockNew and
BlockValid notifications when a new node started up on an existing
network.

Signed-off-by: Logan Seeley <seeley@bitwise.io>