Update how registered consensus engines are stored
Refactors how registered consensus engines are tracked. Previously, they
were stored in the ConsensusNotifier and only used to check if a
consensus engine had been registered before attempting to send messages
to it. This update changes the following:

1. Saves the list as a concurrent map where the key is the connection id
of the registered engine, and the value is a list containing two
elements: the name and version of the engine.
2. Stores the list of registered engines in the core server so it can be
shared between the ConsensusNotifier and the Proxy.
3. The ConsensusRegisterHandler uses the Proxy to add to the list of
registered engines instead of the notifier; it does this by passing the
name, version, and connection id for the engine in its call the proxy's
`register` method.
4. Removes code that is obviated by the above changes.

Signed-off-by: Logan Seeley <seeley@bitwise.io>