Sign and verify consensus messages
Updates how the validator sends and receives consensus peer messages.
Previously, all messages were unsigned and unverified. This update makes
the following changes:

1. Removes GossipConsensusMessage and corresponding network handler in
favor of using GossipMessage for consensus messages; this reduces
duplication and allows for reuse of the existing signature verification
functionality used for GossipMessages.
2. Updates consensus messages so they are sent with a header signature
that is verified when the message is received.

Signed-off-by: Logan Seeley <seeley@bitwise.io>