Update how registered consensus engines are stored
Refactors how the registered consensus engine is tracked. Previously, a
list of registered engines was stored in the ConsensusNotifier and only
used to check if a consensus engine had been registered before
attempting to send messages to it. This update changes the following:

1. Only allows one consensus engine to be registered; this will be used
to enforce the on-chain consensus algorithm setting.
2. Saves the registered consensus engine using the new ConsensusRegistry
class.
3. Stores the registered engine data in the core server so it can be
shared between the ConsensusNotifier and the Proxy.
4. The ConsensusRegisterHandler uses the Proxy to update the registered
consensus engine instead of the notifier; it does this by passing the
name, version, and connection id for the engine in its call the proxy's
`register` method.
5. Removes code that is obviated by the above changes.

Signed-off-by: Logan Seeley <seeley@bitwise.io>