Notify consensus engines when (de)activated
Adds messages that are used to signal the consensus engine when it has
been made the active engine or when it has been deactivated. This will
allow the engines to decide what to do when they are not active (i.e. do
not receive updates from and cannot issue commands to the validator).

Signed-off-by: Logan Seeley <seeley@bitwise.io>