Rename sawtooth-devmode-rust to sawtooth-devmode-engine-rust
This is part of an effort to standardize consensus engine naming. The packages
and published docker images will be named sawtooth-XXXX-engine, and the
binaries will be XXXX-engine.

Devmode-rust is an outlier with the language name included because it is
provided as an example and may be re-implemented in other languages.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>