Add PBFT steps to PoET sim procs in SysAdmin Guide
Add "either/or" choice for PBFT and PoET to all consensus-related steps
and information.

Add PBFT steps/notes in several procedutes to clarify that you must configure
and start at least 4 nodes, and that the genesis block must contain the
validator keys for the other 3 nodes.

Make minor rewording and formatting changes throughout, in order to simplify
the additional layer of complexity for each step.

Signed-off-by: Anne Chenette <chenette@bitwise.io>