Use checked subtraction in queue limit
Ensure that checked subtraction is used when updating the queue limit.
There are scenarios (particularly a race condition that can occur during
genesis block creation) where the current queue limit is less than the
consumed number of batches. This race condition would result in a
"panicked at 'attempt to subtract with overflow'" error.  This change
fixes this issue.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>