Fix container name handling for docker inspect
Call to docker inspect was causing the script to fail because the
container names built by the script did not contain the hash
suffix dynamically generated by docker-compose.

For example:
    latest_test-tp-settings_1

Instead of:
    latest_test-tp-settings_1_47b754af16ac

Modified script to capture and process stdout from subprocess.run
to generate correct container names for docker inspect. Also, modified
script to make docker inspect exit status of 137 equal to 0 for
the purpose of test success determination.

Signed-off-by: Matt Berkau <37899427+mberkau@users.noreply.github.com>