Handle KeyError for missing blocks
Changes to the underlying stores, a KeyError is raised if a block is
missing.  This can occur more aggressively with the BlockManager, which
may remove blocks once they are completely unref'ed. This commit handles
the cases where the unref'ing may have occurred during an ongoing python
operation (due to there being real, Rust threads), and the block and
its resources have been dropped.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>