Handle SetLoggerError gracefully
The SetLoggerError only occurs if `set_logger` has been called more than
once.  This can occur due to the current rust-python-rust nature of the
Sawtooth validator, but may be safely ignored.

A warning message will be logged in this case, in order to provide
notice while it continues to be an issue in some installations.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>