Enable IPv6 support in the validator
This patch will enable the validator to use ZMQ IPv6 sockets.

The ZMQ IPv6 flag will be enabled on a zmq.ROUTER socket if
the binding URL contains an explicit IPv6 address.

The IPv6 flag will enabled on a zmq.DEALER socket if the
binding URL contains either:
- An explicit IPv6 address.
- A hostname that resolves to an IPv6 address.

If the binding URL contains a network interface name, in no
case will the IPv6 flag be enabled.

Signed-off-by: William Katsak <bill@taekion.com>