Filter out None in consensus startup info's peers
When the validator accepts a registration from a consensus engine, it
will return a list of its connected peers in a StartupState object.

This fixes a bug where, if the list of peers contained a None value,
registration would fail when it attempted to encode the public key as
hex.

Signed-off-by: Logan Seeley <seeley@bitwise.io>