Fix ref counting using owned struct BlockRef
Fixes the stability of block (un)ref'ing by using an owned `BlockRef`
struct to represent a particular external reference. This struct
implements the `Drop` trait to automatically decrement the external ref
count for the block when it goes out of scope.

This struct is used for Ref-C (owned by the chain controller) and Ref-D
(owned by the block publisher), but not Ref-B (owned by the completer,
which passes it off to the chain controller/fork cache) since the
completer is implemented in Python. Ref-C was the culprit of the bug
that this commit targets, since it is impacted by an external process
(specifically, the active consensus engine).

Signed-off-by: Logan Seeley <seeley@bitwise.io>