Remove completer from genesis controller
The usage of the completer in the genesis controller was a way of
priming the pump for the chain controller.  When the chain controller
was in python, this worked, as its queues were able to accept blocks
before it was started.  With the rewrite in Rust, the chain controller
no longer is ready to accept blocks until it has started, so the genesis
controller writes the values to the block manager directly.  The chain
controller checks this location on start, and proceeds as normal.  As a
result, the completer can be removed from the genesis controller

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>