Handle multiple commands for same block
This commit updates the chain controller to handle situations where the
consensus engine attempts to issue a commit/ignore/fail command more
than once for the same block. The chain controller will ignore all but
the first command for a block and report an error when this happens.

This handling is enabled by changing the `chain_head` field of the
`ChainControllerState` to hold onto the chain controller's `BlockRef`
(Ref-C) rather than a `Block`. With this change, it is easy to determine
if a block has already been decided on by consensus: if the `BlockRef`
for the block in question is still in the state's `block_references`
set, it has not been decided on; if it is not in the `block_references`,
it has already been decided on, since the ref is simply dropped on an
ignore or fail and moved into the `chain_head` on commit.

A bonus result of this change is that it fixes an issue where Ref-C did
not get created for the genesis block. This commit updates the chain
controller to create the ref in `set_genesis` and
`initialize_chain_head` and store it in the `chain_head` field.

Signed-off-by: Logan Seeley <seeley@bitwise.io>