Remove futures from _SendReceive after a certain timeout.
Some futures produced during peering attempts were never resolved and
_SendReceive did not clean them up in any way. On a longer time scale,
this was leading to excessive memory usage by the validator service.

This commit adds the possibility to set a timeout for a Future object.
If a timeout was set for a Future, FutureCollection is able to clean
that future after a timeout when the remove_expired method is called.
This method is called periodically by _SendReceive (the timeout is equal
to connection_timeout).

Also, futures are explicitly removed when stopping an
OutboundConnection.

Signed-off-by: Yevhenii Babichenko <babichenko@482.solutions>