Remove NewBlockSync on engine registration
Remove the ConsensusRegisterBlockNewSyncHandler which follows an engine
registration.  This sends blocks that are in the block manager, but may
not yet be validated, which violates the current guarantees of the
consensus API where all blocks sent to the consensus engine have been
validated (this is due to constraints placed on the system by PoET 1.0).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>