Allow re-registering consensus engines
To support a consensus engine restarting independently of the validator,
the consensus registry must be able to register a conensus engine again.
When this happens, the engine will register with the same name and
version, but a different connection ID. The registry will simply remove
any old entries associated with the engine name and version before
adding the engine to the registry.

Signed-off-by: Logan Seeley <seeley@bitwise.io>