Make consensus activation atomic
Adds a check to the `ConsensusRegistry.activate_engine` method to check
if the specified engine is already the active engine; if it is, it will
return an `EngineAlreadyActive` exception. This allows for atomically
checking if an engine is already activated and activating it if it's
not. Without this change, there is a race condition caused by checking
the active engine and setting it later.

Signed-off-by: Logan Seeley <seeley@bitwise.io>