Raise and handle EngineNotRegistered
Raise an EngineNotRegistered exception when attempting to activate a
consensus engine that has not been registered.  This will inform the
callers not to do any following activities for the engine (like sending
it a notification) if it hasn't yet connected.

Fixes an issue where duplicate activations were being sent to the engine
due to the notifier queuing messages when there was no active engine.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>