Arch doc: Txn + batch sig must be low-s normalized
Add note with information on low-s normalization for the transaction
and batch headers, along with info that the python3 SDK and others do
this for you, but some libraries don't.

Also changed the existing note to a tip so it'll be visually distinct
(different color).

This doc change addresses the issue reported by STL-1500.

Signed-off-by: Anne Chenette <chenette@bitwise.io>