Make sawadm genesis required settings optional
Adds the `--ignore-required-settings` flag to the `sawadm genesis`
command that allows bypassing the required settings check. Because the
sawtooth settings transaction family is a reference implementation, it's
possible that a different settings transaction family could be used to
set the required settings at genesis. This bypass allows `sawadm
genesis` to pass in this situation. However, the genesis chain
controller will still fail if the required settings aren't in state
as a result of the genesis block, so the same requirements are enforced.

Signed-off-by: Logan Seeley <seeley@bitwise.io>