Avoid duplicate peering request from same endpoint
Example Scenario: If there are three validators A, B and C.
These are configured such that A has B and C as peers,
B has C and A as peers, C has A and B as peers.
Every node initiates an outbound connection to the nodes
configured as their peers.

A peering request is sent when connection is established
successfully. With the above configuration the nodes may have
duplicate peers.

The default number of maximum number of peers is set to 10.
With above configuration, the limit may exceed soon with more
number of peers. Most of which are duplicate peering connections.

Signed-off-by: S m, Aruna <aruna.s.m@intel.com>