Fix SettingsView leak caused by LRU cache usage
Fixes a leak where SettingsViews were never being garbage collected.
This was caused by a SettingsViews' LRU cache entries for the
get_settings method containing a reference to self; this meant that even
when a SettingsView was no longer referenced externally, it would still
not be garbage collected because it internally referenced itself.

A new function decorator, `memoized_method`, is used for the
`SettingsView.get_settings` method. This decorator wraps the `lru_cache`
decorator and uses a weak reference to allow garbage collecting the
settings view. This decorator can be used to memoize any instance
method.

Signed-off-by: Logan Seeley <seeley@bitwise.io>