Don't add batch to block if batch already in queue
Updates the pending batches queue to behave like a HashSet when an item
is appended: it will return true if the item did not previously exist,
false otherwise.

Also updates the block publisher's on_batch_received method to check the
result of pending the batch to the queue: if the batch is already in the
queue, it will not add the batch to the candidate block.

Signed-off-by: Logan Seeley <seeley@bitwise.io>