Guard all modifications to the future collection
Guard all the modifications to the future collection with the lock.
This includes adding the remaining modifications via public methods.

The remove method is not longer used externally, and therefore has been
made a private method, guarded by the set_result function's lock access.

The get method does not require the lock, in that it will catch the
missing value exception.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>