Lazily get chain head root hash in SettingsCache
Updates the SettingsCache to lazily retrieve the chain head state root
hash. Previously, whenever the `get_setting` method was called, we had
to retrieve the state root, even if it wasn't needed due to the setting
already being in the cache.

To accomodate this change in the block publisher's use of the settings
cache (which crosses the FFI boundary), it is no longer given the cache
itself, but rather a closure that wraps the `get_settings` call and the
`block_store.chain_head_state_root` function.

Signed-off-by: Logan Seeley <seeley@bitwise.io>