Fix SettingsCache settings existence check
Fixes how the SettingsCache checks if a setting has already been
retrived. Previously, the cache assumed that if the value in the cache
was None, the setting was never retrieved; instead, if the key is in the
cache, the setting has already been retrieved (even if the value is
None), so we don't need to read from state.

Signed-off-by: Logan Seeley <seeley@bitwise.io>