Fix IdentityCache existence check
Fixes how the IdentityCache checks if an item has already been
retrieved. Previously, the cache assumed that if the value in the cache
was None, the item was never retrieved; instead, if the item is in the
cache, the item has already been retrieved (even if the value is
None), so we don't need to read from state.

Signed-off-by: Logan Seeley <seeley@bitwise.io>