Lazily get chain head root hash in IdentityCache
Updates the IdentityCache to lazily retrieve the chain head state root
hash. Previously, whenever the ``get_role` or `get_policy` methods were
called, we had to retrieve the state root, even if it wasn't needed due
to the requested value already being in the cache.

To make this change possible, we also had to slightly modify the
PermissionVerifier to keep track of whether or not a chain head exists.

Signed-off-by: Logan Seeley <seeley@bitwise.io>