Remove the dependency on python SDK
Remove the dependency on the python SDK to avoid the collisions of the
generated protobuf classes from the essentially the same source.

This replaces the use of the IntkeyMessageFactory with direct creation
of intkey transactions using the protobuf classes in the
`sawtooth_validator` package.

Note: the events and receipts test still need the sdk (currently) for
use of the Stream class.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>