Allow additional protocols for consensus engines
Adds an (optional) list of additional protocols that a consensus engine
supports as part of the registration process. These additional protocols
can be used for backwards compatiblity. If the values of the
`sawtooth.consensus.algorithm.name` and
`sawtooth.consensus.algorithm.version` settings match the name/version
of an engine or the name/version pair is in the engine's list of
additional protocols, that engine will be activated.

More than one engine can support a given protocol; if this is the case,
the engine that registered first will be activated if that protocol
becomes active.

If an engine registers and it supports all of the same protocols as an
existing registered engine, the existing engine is unregistered (and
deactivated if it was active).

Signed-off-by: Logan Seeley <seeley@bitwise.io>