Fix wait certificate nonce generation
Wait certificate creation is protected in part by using the monotonic
counter to enforce singleton behavior across possible instances of the
enclave. This fix makes the counter use more atomic to complicate
reentrant attacks.

As further protection this patch replaces the use of an rng with a
hash so as to generate an unpredictable but deterministic nonce.

An alternate implementation would use hmac which is only available in
later versions of the SGX SDK than is used by PoET at this time.

Signed-off-by: Dan Middleton <dan.middleton@intel.com>