Collect futures for explicit cancellation
Using all_tasks to cancel tasks is dangerous, as it can result in a
concurrent modification of the list of tasks.  The prescribed approach
is to collect the tasks that may be cancelled in a list, and iterate
over those tasks when cancelling.  This will avoid cancelling the
cancel task itself (and other shutdown tasks).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>