Add per instance lru cache methods for StateView
This adds an instance-level lru cache for the StateView methods
get(), addresses(), and leaves(). Since StateView is a non-mutable
read interface to a specific version of the underlying merkle
database identified by a state root hash, these calls are
guaranteed to return the same results on every invocation. This
behavior makes them good candidates for per instance caching, and
substantially reduces the number of underlying database get calls.

Signed-off-by: James Mitchell <mitchell@bitwise.io>