Make GossipDuplicateHandler use a local cache vs. a completer check
In a gossip network, often multiple duplicate messages arrive from
peers in short order. Previously, the duplicate handler was performing
an existence check (has_block, has_batch) against the completer.
Since the completer likely does not process the first message by
the time the last duplicate is seen, the completer's "caches" were
not filled, resulting in a high volume of block store calls (resulting
in "not in" results to the duplicate handler). This change moves to
an LRU cache of message signature headers managed by the handler itself.
This should allow the duplicate handler to behave as expected (dropping
short-term duplicates) without excessive existence check calls.

Signed-off-by: James Mitchell <mitchell@bitwise.io>