Make GossipBroadcastHandler use a local cache vs. a completer check
In sawtooth, message broadcasting occurs prior to the message
being passed to the completer. Previously, the broadcast handler
was performing an existence check against the completer to determine
if re-broadcasting was required. This resulted in a completer
"cache" check and a block store existence check for an initial
message, despite the fact that the completer could not have
handled the message yet (as it comes after broadcast in the
dispatch list). To exacerbate matters, if duplicate messages
are passed to the broadcast handler in short order, the completer
checks would still return "not in", and the messages would be
broadcast multiple times to peers. This change moves to an LRU
cache of message header signatures managed by the handler itself.

Signed-off-by: James Mitchell <mitchell@bitwise.io>