Replace heartbeat with reap argument in Interconnect
Heartbeat is now enabled for all zmq connections, however
to remain backword compatible, connections cannot be removed if
they do not respond to a PingRequest. Instead, if the argument is
reap is set to True connections that do not respond will eventually
be removed. If reap is False, the ping will still be sent but
no response is expected/required and the connetion will not be
removed.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>