Add LmdbOrderedStoreIter
Add the `LmdbOrderedStoreIter` struct that is returned by the
`LmdbOrderedStore.iter()` method. This iterator loads items into an
in-memory cache as needed and provides an iterator over the values in
the ordered store. Additionally, it provides the option to specify a
range to get only a subset of items in the database.

Signed-off-by: Logan Seeley <seeley@bitwise.io>