Fix timestamp check in block_info for catch up
This is similar to 5e7315a9f0e3c8863327034c036b28e70850112c, which was
a fix for the python version of block_info. To quote that commit:

"The timecheck needs to ensure that the timestamp is only ahead of
a transaction processor's local time by the value of tolerance. The use
of absolute value enforced that this time check is within tolerence of
the local time. This fails validation in the case where a node is
catching up on a chain the may contain blocks that were published more
than time-tolerence in the past.  Correcting this to ensure the the
timestamp is no more than tolerence in the future, ensures that the
transaction can still be validated during a catch up scenario."

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>