Replace atomic bool in publisher with message enum
Replaces the `Exit` struct in the block publisher with an updated
channel that is used to shutdown the publisher thread. This eliminates
the need to use `recv_timeout` on the receiver, which can cause panics
in some situations.

Also removes the `IncomingBatchReceiver` struct and replaces it with a
plain `Receiver`, since the `IncomingBatchReceiver` is unnecessary (it
does not need to add any seen batch IDs, since they'll already have been
added by the sender).

Signed-off-by: Logan Seeley <seeley@bitwise.io>