Check if check_connections is None before calling
This check was called when the monitor thread was notified
of a disconnection and would check all existing external
transaction processors. With the transact integration and
removal of the python executor, the check_connection
function was never set.

This commit adds a check to skip calling check_connections
if it is none.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>