Use provided block in BatchTracker.chain_update
Updates the `chain_update` method for the batch tracker to use the
provided block for notifying of committed batches rather than checking
the commit store. This fixes a bug that was created by a recent change
to the chain controller; the chain controller now notifies of committed
batches before persisting the committed block to the commit store. This
resulted in the batch tracker not properly notifying of committed
batches.

The fix is to not check the commit store in `chain_update`, but rather
just use the batch IDs in the provided block. This makes the batch
tracker resilient to receiving the "commit" notification from the chain
controller before the chain controller has persisted all of its state.
In addition to fixing the bug, this avoids an unnecessary database call.

Signed-off-by: Logan Seeley <seeley@bitwise.io>