Ignore two broken lmdb tests in adm
These tests look for an lmdb database and error that it is not found. It
is not clear if additional setup is expected for these tests or if it
indicates a behavioral change in lmdb auto-creation. In either case,
perhaps these tests should be restored in the future in a non-broken
form.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>